+++ app/src/main/java/com/example/melitest/data/model/PagingModel.kt
@@ -3,8 +3,8 @@ package com.example.melitest.data.model
 import com.google.gson.annotations.SerializedName
 
 data class PagingModel(
-    @SerializedName("total") var total: Long? = null,
-    @SerializedName("primary_results") var primaryResults: Long? = null,
-    @SerializedName("offset") var offset: Long? = null,
-    @SerializedName("limit") var limit: Long? = null
+    @SerializedName("total") var total: Float? = null,
+    @SerializedName("primary_results") var primaryResults: Float? = null,
+    @SerializedName("offset") var offset: Float? = null,
+    @SerializedName("limit") var limit: Float? = null
 )
