+++ app/src/main/java/com/example/melitest/data/model/RatingsModel.kt
@@ -3,7 +3,7 @@ package com.example.melitest.data.model
 import com.google.gson.annotations.SerializedName
 
 data class RatingsModel(
-    @SerializedName("negative") var negative: Double? = null,
-    @SerializedName("neutral") var neutral: Double? = null,
-    @SerializedName("positive") var positive: Double? = null
+    @SerializedName("negative") var negative: Float? = null,
+    @SerializedName("neutral") var neutral: Float? = null,
+    @SerializedName("positive") var positive: Float? = null
 )
