+++ app/src/main/java/com/example/melitest/data/model/SalesModel.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class SalesModel(
 
     @SerializedName("period") var period: String? = null,
-    @SerializedName("completed") var completed: Long? = null
+    @SerializedName("completed") var completed: Float? = null
 
 )
