+++ app/src/main/java/com/example/melitest/data/model/TransactionsModel.kt
@@ -4,9 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class TransactionsModel(
 
-    @SerializedName("canceled") var canceled: Long? = null,
-    @SerializedName("completed") var completed: Long? = null,
+    @SerializedName("canceled") var canceled: Float? = null,
+    @SerializedName("completed") var completed: Float? = null,
     @SerializedName("period") var period: String? = null,
     @SerializedName("ratings") var ratings: RatingsModel? = RatingsModel(),
-    @SerializedName("total") var total: Long? = null
+    @SerializedName("total") var total: Float? = null
 )
