+++ app/src/main/java/com/example/melitest/data/model/ValueStructModel.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ValueStructModel(
 
-    @SerializedName("number") var number: Double? = null,
+    @SerializedName("number") var number: Float? = null,
     @SerializedName("unit") var unit: String? = null
 
 )
