+++ app/src/main/java/com/example/melitest/data/model/ValuesModel.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class ValuesModel(
     @SerializedName("id") var id: String? = null,
     @SerializedName("name") var name: String? = null,
-    @SerializedName("results") var results: Long? = null
+    @SerializedName("results") var results: Float? = null
 )
